Implement @locals and @local++/--
Parser now processes assignment statements and usages of @local
variables.  @local variables may not be used before assignment.

@local variables can now beincremented or decremented - but note that
these are STATEMENTS, not expressions.  You can't do @b++ / 2.

The interpreter reads the first byte of the bytecode and uses this
to allocate space on the stack for all the locals used int the item.

New opcodes:
c <index> - Saves the top of the stack into the indexed local
            (stack is reduced by one)
e <index> - Grabs a copy of the value of the indexed local and
            pushes it onto the top of the stack.
f <index> - increment the int value at <index>
g <index> - decrement the int value at <index>

The interpreter now returns the top of the stack, or if the stack is
empty, it returns VALUE_NIL.

Don't include locals in stack size: although locals are technically
on the stack they are in fact degenerate freeloaders.  Do not report
them via size_stack().