Implement WHILE-DO loop
This involves a bit of rewiring of the parser, to pass more state.  The
lexer has also been reconfigured to be case-insensitive.

Fixed a silly (and as-yet untriggered) bug where 64-bit integer values were
being emitted as an unsigned int - they should be a signed int.

New opcode 'j': unconditional jump.  Interpret the next two bytes as a
signed short, and jump forward or backward in the bytecode by that amount.
New opcode 'k': jump-if-false.  As opcode 'j', but first pop the stack
and only jump if it evaluates to false.