Revert "Implement infinite loop trap"
This reverts commit e626ed3b70d1f061b9641f23f0d02312e6dfaf1a.

This won't work, because items call other items, which can create an infinite
loop in a way which is not obvious and would get around this this trap.

Will need to be re-done using alarm(), which gets reset at the top of every
game loop, probably after processing player input (the putative io.input()
pseudoitem).  Don't worry about this until after items are fully implemented
and there is an io.input() to fiddle with.