Add CODE statement
Add the code () assignment statement.  This permits the following
language construction:
    foo = code ( bar; );
When the compiler encounters this, it will gather up all the code between
the ( and ) after the code keyword, and pass that to a new compiler
instance.  Assuming successful compilation, it will then assign the
resultant bytecode to the item 'foo'.  Thereafter, calling foo will result
in the associated bytecode being executed, and its return value being the
resultant value of foo.

UNTESTED
WIP WIP WIP

Changes:
- Teach the lexer about the code () construct.
- Teach the parser to output correct bytecode.
- Teach the disassembler to disassemble the B opcode.
- Skeleton opcode handler for the B opcode.

To do:
- Integrate into the interpreter (which will involve integrating the
  compiler).