Implement return keyword.
It must be possible to stop execution of an item before it reaches the end.
The return keyword does this.  It does not take a value, it simply inserts
a 'h' (halt) opcode into the bytecode, so that the interpreter stops
interpreting at that point and returns to the calling item.  The value
on top of the stack is the value returned to the calling item (or nil if
there is no value on the stack).