Prepare for parameters
Code items need to be able to receive parameters.  This commit prepares
the way for that work.

Code can already contain local variables, and parameters are just a
special type of local variable.  Thus, although there is a separate
parameter count, its total is included in the local count as well.
Consider, in a future timeline, an item with this code:

test = code {@a, @b} ( @c = @b * 10; @a + @c; );

The parameter count will be 2.  The local count will be 3 (parameters
@a and @b, and local @c).

In the bytecode:
  Byte 0 is the total number of locals.
  Byte 1 is the number of locals which are parameters.