Implement item parameter definition
The compiler now recognises and emits bytecode for parameters in
item definitions.  Parameters are optional, and if present must
be a comma-separated list of local variables, enclosed within braces thus:
  mult = code {@a, @b} ( @a * @b; );
Parameter definition in the bytecode is signalled by opcode P, and comes
immediately after the B opcode (if it comes at all) and immediately before
the sourcecode.

The interpreter now properly handles parameters in item definitions when
doing a nested compilation.

None of this can be meaningfully tested until item arguments are a thing,
and that is the subject of the next commit.