Move item_out buffer from LOCAL_t to SCANNER_STATE_t
There is a problem with passing items as arguments to other items.  This is
(at least partially) because there is only one item_out buffer.  To fix
this, move the item_out buffer into the scanner state struct, in
preparation for turning it into an array of buffers (akin to how
control structures are nested).