Library calls implementation
Library calls look like items, but they aren't.  They are, in a sense,
"pseudoitems", and behave in the same way.  A library call is always
of the form:
  libname.callname{args}
Currently, the library names are hard-coded into src/lexer.l but this
can probably be improved.

A new opcode, A, is introduced, which instructs the interpreter to
lookup and execute a library call.

Library calls must always push a return value onto the stack, even if
that is nil.

op_libcall() is just a dummy at the moment, and there are no actual
library functions in src/libcalls.c, but the infrastructure is there.