Remove slab allocator
Something subtly wrong with it.  Removed because it is too early to think
about optimising, and I don't want to spend time fixing something obscure
when it might not be necessary.

Kept the allocator API, but they are just wrappers to GROW_ARRAY and
FREE_ARRAY for now.

This solves a weird memory corruption bug which defied identification.