Start implementing the task subsystem
Tasks are the core of the game - the runloop listens for activity on a
task, whether that is network activity or timer expiry, and then runs
the item associated with the task.  From within the game, tasks can be
dynamically added and removed - these are timer tasks.

Basic implementation of task.newgametask, but no way of stopping it yet.