Implement net.input libcall
This is the library call which tells the game whether there has been
any activity (connections, disconnections or data).  It is intended to
be called only by the item which is set as the `input` task.

The input task detaults to `input`, but can be changed by a commandline
option.

All activity is processed, but currently no data is handed to the input
task for processing (it just sits in the connection input buffer).

Lightly tested.  Not guaranteed to not explode.