Add nthname{@item, @index}
This is a slow way of iterating through the children of an item.  Faster
ways are possible, but that is a job for the optimisation phase.

@item is an expression which evaluates to a string.  This string is looked
up in the itemstore to see if it is an item.  If it is, index is evaluated
to see if it comes out as an integer >= 0.  If it is, the name of the
indexth child of the item is returned.  If either of the arguments are
invalid, or the index is out of range, nil is returned.